Move PatchEvent(s) and add PluginEvent(s)
